Skeleton of api POST request json
This will need to be finished as we gain functionality in the item
class to do so, however, pretty much every type of item in the game
can be searched by this method. There are comments throughout the
get_json() function which explain cases for different types
of items we can request.

To finish this is really just ifing out the type of item and changing
what we fill depending on what we have. Most of the differences
here are involved with the name/type fields.

Signed-off-by: Kevin Morris <kevr.gtalk@gmail.com>